made ConfirmDeleteButton a reusable component
