+++ components/buttons/index.tsx
@@ -1,0 +1,3 @@
+import ConfirmDeleteButton from './ConfirmDeleteButton'
+
+export { ConfirmDeleteButton }
